+++ src/main/java/de/sakurajin/sakuralib/mixin/minecraft/WorldEntryMixin.java
@@ -1,5 +1,6 @@
 package de.sakurajin.sakuralib.mixin.minecraft;
 
+import de.sakurajin.sakuralib.SakuraLib;
 import de.sakurajin.sakuralib.internal.SakuraLibConfig;
 import de.sakurajin.sakuralib.internal.ui.modUpdateScreen;
 import net.minecraft.client.MinecraftClient;
@@ -12,7 +13,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(value = net.minecraft.client.gui.screen.world.WorldListWidget.WorldEntry.class, remap = false)
+@Mixin(value = net.minecraft.client.gui.screen.world.WorldListWidget.WorldEntry.class)
 public abstract class WorldEntryMixin {
     @Shadow @Final private MinecraftClient client;
     @Shadow @Final public LevelSummary level;
@@ -20,7 +21,7 @@ public abstract class WorldEntryMixin {
     @Inject(method = "play", at = @At("HEAD"), cancellable = true)
     public void sakuralib$play(CallbackInfo ci){
         //check if the warning screen is enabled
-        if(!SakuraLibConfig.createAndLoad().ENABLE_WARNING_SCREEN()){
+        if(!SakuraLib.CONFIG.ENABLE_WARNING_SCREEN()){
             return;
         }
         if(!this.level.isUnavailable()){